add: [core] Behavior correction for automatic test database instance launcher
