+++ common_tests/test_template.py
@@ -1,12 +1,13 @@
 # -*- coding: utf-8 -*-
 
-import sys
-import os
 import logging
+import os
+import sys
+
 sys.path.append(os.path.abspath(os.path.pardir))
-from carlhauser_server_tests.context import *
 import pathlib
 import unittest
+
 
 class test_template(unittest.TestCase):
     """Basic test cases."""
